+++ app/src/main/java/org/mozilla/social/MainApplication.kt
@@ -14,9 +14,8 @@ import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
 import org.koin.dsl.module
 import org.mozilla.social.common.commonModule
-import org.mozilla.social.core.analytics.glean.Analytics
-import org.mozilla.social.core.analytics.glean.GleanAnalytics
-import org.mozilla.social.core.analytics.glean.analyticsModule
+import org.mozilla.social.core.analytics.Analytics
+import org.mozilla.social.core.analytics.analyticsModule
 import org.mozilla.social.core.data.AuthCredentialObserver
 import org.mozilla.social.core.data.repositoryModule
 import org.mozilla.social.core.database.databaseModule
@@ -37,7 +36,7 @@ class MainApplication : Application(), ImageLoaderFactory {
 
     private lateinit var authCredentialObserver: AuthCredentialObserver
 
-    val analytics: Analytics by inject()
+    private val analytics: Analytics by inject()
 
     override fun onCreate() {
         super.onCreate()
