+++ core/analytics/src/main/java/org/mozilla/social/core/analytics/AnalyticsModule.kt
@@ -1,0 +1,12 @@
+package org.mozilla.social.core.analytics
+
+import org.koin.dsl.module
+import org.mozilla.social.core.analytics.glean.GleanAnalytics
+
+val analyticsModule = module {
+    single<Analytics> {
+        object : Analytics {}
+        // use glean when we have setting ready
+//        GleanAnalytics()
+    }
+}
