--- core/analytics/src/main/java/org/mozilla/social/core/analytics/glean/AnalyticsModule.kt
@@ -1,7 +1,0 @@
-package org.mozilla.social.core.analytics.glean
-
-import org.koin.dsl.module
-
-val analyticsModule = module {
-    single<Analytics> { GleanAnalytics() }
-}
