+++ core/analytics/src/main/java/org/mozilla/social/core/analytics/glean/GleanAnalytics.kt
@@ -3,61 +3,23 @@ package org.mozilla.social.core.analytics.glean
 import android.content.Context
 import mozilla.telemetry.glean.BuildInfo
 import mozilla.telemetry.glean.Glean
+import org.mozilla.social.core.analytics.Analytics
 import org.mozilla.social.core.analytics.GleanMetrics.Identifiers
 import org.mozilla.social.core.analytics.GleanMetrics.Ui
 import java.util.Calendar
 
-interface Analytics {
-    fun initialize(context: Context)
-
-    fun uiEngagement(
-        engagementType: String? = null,
-        engagementValue: String? = null,
-        mastodonAccountHandle: String? = null,
-        mastodonAccountId: String? = null,
-        recommendationId: String? = null,
-        uiAdditionalDetail: String? = null,
-        uiIdentifier: String? = null
-    )
-
-    fun uiImpression(
-        mastodonAccountHandle: String? = null,
-        mastodonAccountId: String? = null,
-        mastodonStatusId: String? = null,
-        recommendationId: String? = null,
-        uiAdditionalDetail: String? = null,
-        uiIdentifier: String? = null
-    )
-
-    fun setAdjustDeviceId(adjustDeviceId: String)
-
-    fun setFxaAccountId(fxaAccountId: String)
-
-    fun setMastodonAccountHandle(mastodonAccountHandle: String)
-
-
-    fun setMastodonAccountId(mastodonAccountId: String)
-
-    fun setUserAgent(userAgent: String)
-
-    fun clearLoggedInIdentifiers()
-}
-
 class GleanAnalytics : Analytics {
     override fun initialize(context: Context) {
-//
-//        This will be re-added when settings for tracking is done
-//
-//        val buildInfo = BuildInfo("1", "1", Calendar.getInstance())
-//
-//        Glean.setLogPings(true)
-//        Glean.setDebugViewTag("moso-android-debug")
-//
-//        Glean.initialize(
-//            applicationContext = context,
-//            uploadEnabled = true,
-//            buildInfo = buildInfo
-//        )
+        val buildInfo = BuildInfo("1", "1", Calendar.getInstance())
+
+        Glean.setLogPings(true)
+        Glean.setDebugViewTag("moso-android-debug")
+
+        Glean.initialize(
+            applicationContext = context,
+            uploadEnabled = true,
+            buildInfo = buildInfo
+        )
     }
 
     override fun uiEngagement(
