+++ feature/account/src/main/java/org/mozilla/social/feature/account/AccountViewModel.kt
@@ -20,8 +20,7 @@ import org.mozilla.social.common.Resource
 import org.mozilla.social.common.logging.Log
 import org.mozilla.social.common.utils.StringFactory
 import org.mozilla.social.common.utils.edit
-import org.mozilla.social.core.analytics.glean.Analytics
-import org.mozilla.social.core.analytics.glean.GleanAnalytics
+import org.mozilla.social.core.analytics.Analytics
 import org.mozilla.social.core.data.repository.AccountRepository
 import org.mozilla.social.core.data.repository.StatusRepository
 import org.mozilla.social.core.data.repository.model.status.toExternalModel
@@ -33,7 +32,6 @@ import org.mozilla.social.core.ui.R
 import org.mozilla.social.core.ui.postcard.PostCardDelegate
 import org.mozilla.social.core.ui.postcard.PostCardNavigation
 import org.mozilla.social.core.ui.postcard.toPostCardUiState
-import org.mozilla.social.model.Account
 import timber.log.Timber
 
 class AccountViewModel(
